blob: Fix spdk_bs_blob_decouple_parent when blob's ancestor is an esnap.
This used to fail with `-ENOENT` because `bs_inflate_blob_done` didn't
differentiate between blob-backed parents and esnap-backed parents, and
tried to open a blob with id `_parent->parent_id`, which was
`SPDK_BLOBID_EXTERNAL_SNAPSHOT`.

In such cases, the inflated blob will need to use the same back device
of its ex-parent. Since lifetime of back devices are same as the blob
owning it, we need to clone the back device. To allow this, a new
`clone` function pointer is introduced in `struct spdk_bs_dev`.

A future change will implement it for `struct blob_bdev`.

Change-Id: I31de50c8b5a4d50349573c1e1b7918b950b83193
Signed-off-by: Hadi Moshayedi <hadi@ubicloud.com>