blob: add the blob_unref_back_bs_dev helper function.
This is used when a blob doesn't need its current bs_dev anymore. In
current patch we just simply destroy the bs_dev, but in a future patch
we'll extend the logic to the case where there can be multiple refs to
the same bs_dev.

See https://review.spdk.io/gerrit/c/spdk/spdk/+/24247

Change-Id: Ie959b681735373400db6d4c56dddf44f6facc316
Signed-off-by: Hadi Moshayedi <hadi@ubicloud.com>